pyproject.toml
[project]
name = "simple-ml-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]
